change component name from SettingsList to SettingItem
